requirements.txt
fastapi
uvicorn
requests
pandas
pandas-ta
python-dotenv